requirements.txt
python-dotenv>=1.0
requests>=2.31
pandas>=2.2
oandapyV20>=0.6.3
krakenex>=2.1.0
python-telegram-bot==13.15
flask>=3.0
